// src/styles/theme.d.ts
export declare const colors: {
    primary: string;
    secondary: string;
    primaryDark: string;
    yellow: string;
    fuchsia: string;
    teal: string;
    slateblue: string;
    gray: string;
    grayLight: string;
    white: string;
    black: string;
};
export declare const fonts: {
    SansLight: string;
    SansRegular: string;
    SansSemiBold: string;
    SansBold: string;
};
export declare const screenSizesPx: {
    mobile: string;
    tablet: string;
};
export declare const screenSizes: {
    mobile: number;
    tablet: number;
};